% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_pos_exist.R
\name{false_pos_exist}
\alias{false_pos_exist}
\title{Computes false positive rate}
\usage{
false_pos_exist(iteration_num, sim_data)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Computes false positive rate
}
