% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lodder}
\alias{lodder}
\title{Meta analysis of money priming studies}
\format{
A data frame with 287 rows and 4 variables:
\describe{
\item{study}{code identifying the study}
\item{yi}{effect size (what units?)}
\item{vi}{variance}
\item{zi}{z-score}
\item{preregistered}{logical indicating whether study was preregistered}
}
}
\source{
\url{TODO}
}
\usage{
lodder
}
\description{
Longer description
}
\keyword{datasets}
