% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{money_priming_meta}
\alias{money_priming_meta}
\title{Meta-analysis of money priming studies}
\format{
A data frame with 287 rows and 4 variables:
\describe{
\item{study}{Code identifying the study}
\item{yi}{Point estimate on the Hedges' \emph{g} scale}
\item{vi}{Variance of point estimate}
\item{zi}{Z-score}
\item{preregistered}{Logical indicating whether study was preregistered}
}
}
\usage{
money_priming_meta
}
\description{
Dataset from a meta-analysis of experimental studies on the effect of money
primes on a variety of psychological and behavioral outcomes, in which some
studies were preregistered \insertCite{lodder2019}{metabias}.
}
\references{
\insertRef{lodder2019}{metabias}

\insertRef{lodder2020}{metabias}
}
\keyword{datasets}
