% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phacking-package.R
\docType{package}
\name{phacking-package}
\alias{phacking-package}
\alias{phacking}
\title{The 'phacking' package.}
\description{
Methods for meta-analysis corrected for phacking.
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R
package version 2.21.5. https://mc-stan.org
}
