% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{rtma_qqplot}
\alias{rtma_qqplot}
\title{QQ plot for a right-truncated meta-analysis}
\usage{
rtma_qqplot(rtma)
}
\arguments{
\item{rtma}{Output of \code{phacking_rtma()}.}
}
\description{
QQ plot for a right-truncated meta-analysis
}
\examples{
lodder_rtma <- phacking_rtma(lodder$yi, lodder$vi, parallelize = FALSE)
rtma_qqplot(lodder_rtma)
}
