% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_exp_simulations.R
\name{run_exp_simulations}
\alias{run_exp_simulations}
\title{Simulations optional stopping.}
\usage{
run_exp_simulations(sim_params, num_iterations, num_cores, seed, step_size = T)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Simulations optional stopping.
}
