% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{run_mult_simulations}
\alias{run_mult_simulations}
\title{Generates data according to a covariance matrix.}
\usage{
run_mult_simulations(num_iterations, seed, param_list, num_cores = 3)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Generates data according to a covariance matrix.
}
