% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{z_density}
\alias{z_density}
\title{Z-score density plot}
\usage{
z_density(zi, alpha_select = 0.05, crit_color = "red")
}
\arguments{
\item{zi}{Vector of within-study z-scores.}

\item{alpha_select}{Alpha level.}

\item{crit_color}{Color for line and text are critical z-score.}
}
\description{
Z-score density plot
}
\examples{
z_density(lodder$zi)
}
