% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{phacking_rtma}
\alias{phacking_rtma}
\title{Right-truncated meta-analysis}
\usage{
phacking_rtma(
  yi,
  vi,
  sei,
  favor_positive = TRUE,
  alpha_select = 0.05,
  stan_control = list(adapt_delta = 0.98, max_treedepth = 20),
  parallelize = TRUE
)
}
\arguments{
\item{yi}{A vector of point estimates to be meta-analyzed.}

\item{vi}{A vector of estimated variances for the point estimates.}

\item{sei}{A vector of estimated standard errors for the point estimates
(only one of \code{vi} or \code{sei} needs to be specified).}

\item{favor_positive}{\code{TRUE} if phacking is assumed to favor positive
estimates; \code{FALSE} if assumed to favor negative estimates (defaults to
\code{TRUE}).}

\item{alpha_select}{Alpha level that phacking is assumed to operate over.}

\item{stan_control}{List passed to \code{rstan::sampling()} as the
\code{control} argument.}

\item{parallelize}{Logical indicating whether to parallelize sampling.}
}
\value{
A list with three elements, \code{params}, \code{stats} and
\code{fit}.
}
\description{
Right-truncated meta-analysis
}
\examples{
phacking_rtma(lodder$yi, lodder$vi, parallelize = FALSE)
}
