% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phacking.R
\name{phacking_rtma}
\alias{phacking_rtma}
\title{Right-truncated meta-analysis}
\usage{
phacking_rtma(
  yi,
  vi,
  sei,
  favor_positive = TRUE,
  alpha_level = 0.05,
  stan_control = list(adapt_delta = 0.98, max_treedepth = 20)
)
}
\arguments{
\item{yi}{A vector of point estimates to be meta-analyzed.}

\item{vi}{A vector of estimated variances for the point estimates.}

\item{sei}{A vector of estimated standard errors for the point estimates.}

\item{favor_positive}{\code{TRUE} if phacking is assumed to favor positive
estimates; \code{FALSE} if assumed to favor negative estimates (defaults to
\code{TRUE}).}

\item{alpha_level}{Alpha level at which}

\item{stan_control}{List passed to \code{rstan::sampling()} as the
\code{control} argument.}
}
\value{
A list with two elements, \code{stats} and \code{fit}.
}
\description{
Right-truncated meta-analysis
}
